updating import references to relocated subfolder
