+++ api/src/modules/generateMulesoftAPI_attendances_get.js
@@ -1,3 +1,3 @@
-import getConfigurationValueByKey from "./getConfigurationValueByKey";
+import getConfigurationValueByKey from "./dot-env-configuration/getConfigurationValueByKey";
 const generateMulesoftAPIEndpoint_attendances_get = (coachId,teamSeasonId,sessionId) => `${getConfigurationValueByKey("MULESOFT_API_ROOT_URL")}/coach/${coachId}/teamseasons/${teamSeasonId}/sessions/${sessionId}/attendances`;
 export default generateMulesoftAPIEndpoint_attendances_get;
