+++ api/src/modules/generateMulesoftAPI_attendances_post.js
@@ -1,3 +1,3 @@
-import getConfigurationValueByKey from "./getConfigurationValueByKey";
+import getConfigurationValueByKey from "./dot-env-configuration/getConfigurationValueByKey";
 const generateMulesoftAPIEndpoint_attendances_post = () => `${getConfigurationValueByKey("MULESOFT_API_ROOT_URL")}/attendances`;
 export default generateMulesoftAPIEndpoint_attendances_post;
