+++ api/src/modules/generateMulesoftAPI_coach_coachIdAndDate.js
@@ -1,3 +1,3 @@
-import getConfigurationValueByKey from "./getConfigurationValueByKey";
+import getConfigurationValueByKey from "./dot-env-configuration/getConfigurationValueByKey";
 const generateMulesoftAPIEndpoint_coach_coachIdAndDate = (coachId,dateFilter) => `${getConfigurationValueByKey("MULESOFT_API_ROOT_URL")}/coach/${coachId}/all?date=${dateFilter}`;
 export default generateMulesoftAPIEndpoint_coach_coachIdAndDate;
