+++ api/src/modules/generateMulesoftAPI_coach_sessions_post_new.js
@@ -1,3 +1,3 @@
-import getConfigurationValueByKey from "./getConfigurationValueByKey";
+import getConfigurationValueByKey from "./dot-env-configuration/getConfigurationValueByKey";
 const generateMulesoftAPIEndpoint_coach_sessions_post_new = () => `${getConfigurationValueByKey("MULESOFT_API_ROOT_URL")}/sessions`;
 export default generateMulesoftAPIEndpoint_coach_sessions_post_new;
