+++ api/src/modules/generateMulesoftAPI_enrollments_post.js
@@ -1,3 +1,3 @@
-import getConfigurationValueByKey from "./getConfigurationValueByKey";
+import getConfigurationValueByKey from "./dot-env-configuration/getConfigurationValueByKey";
 const generateMulesoftAPIEndpoint_enrollments_post = () => `${getConfigurationValueByKey("MULESOFT_API_ROOT_URL")}/enrollments`;
 export default generateMulesoftAPIEndpoint_enrollments_post;
